mem-ruby: Add lastTouchTick to ReplacementData
This is a step towards merging the replacement policies from Classic to Ruby.
As we need Ruby to inherit from Classic, we have to add getLastAccess() to
BaseReplacementPolicy in Classic. Since lastTouchTick, which is defined in
LRUReplData and MRUReplData, will be used in getLastAccess(), we have to
delete lastTouchTick in LRUReplData and MRUReplData and add lastTouchTick
to their parent, ReplacementData.

Change-Id: I1ef6e829e21779fd516d1072c935409f4cee6898