mem-ruby: Make AbstractReplacementPolicy inherited from BaseReplacementPolicy
This is a step towards merging the replacement policies from Ruby to Classic.
As we need Ruby to inherit from Classic, we have to add getLastAccess() and
useOccupancy() to BaseReplacementPolicy in Classic. The getLastAccess() will
be called differently in CacheMemory depending on which policy is used. The
useOccupancy() will indicate whether it's WeightedLRUPolicy.

Change-Id: I85a807b5ff52ab58ddcbc12c1fe8074119171a9b