mem: Debug compiling error for GPU protocols
After changing the GPU protocols to use AbstractCacheEntry, and
resolving the overriden errors in src/gpu-compute folder, the
compiler complaints that variable m_data_cache_hit_latency is not declared.
Though this file inherits from Sequencer.hh, and Sequencer.hh has
this variable delcared, this error still shows. To eliminate this
error, I added declaration of this variable in this file.

Change-Id: I5c54bd79bce04f1db9b454e99f00d8f8d56d47de