mem-ruby: update lastTouchTick in touch functions for all replacement policies
This is a step towards merging replacement policies in Ruby and Classic.
We add lastTouchTick to ReplacementData structure so that when we use
BaseReplacementPolicy in Ruby system, we can return the correct value for
getLastAccess() function. To do this, we have to update lastTouchTick
in touch functions for all replacement policies, even though some of these
replacement policies previously didn't use lastTouchTick.

Change-Id: Id0aa55186f469d486fba24fa30d5901caec1617a